# Patch/beamconvspeedup