Squeeze some more speed out of the SMILES reader at the Beam->CDK handoff. This might be a bit of a no-op as testing the AtomRef patching (on going) I lose about a second in the SMILES Parser, so the solution is to squeeze an extra second out of the current implementation... thus balance is restored.

### Current
/data/chembl_22.smi	11.063s
/data/chembl_22.smi	10.401s
/data/chembl_22.smi	10.385s
/data/chembl_22.smi	10.385s
/data/chembl_22.smi	10.387s

### Patch
/data/chembl_22.smi	10.021s
/data/chembl_22.smi	9.403s
/data/chembl_22.smi	9.391s
/data/chembl_22.smi	9.371s
/data/chembl_22.smi	9.357s

Times will go back to ~10.5 seconds if I get the AtomRef integrated nicely.